fix db locking

+++ assembly/create.sh
@@ -6,15 +6,20 @@ cd reactive-playground-docker
 ./create.sh $@ || exit 1
 cd ..
 
-docker stop reactive-playground
-docker rm reactive-playground
+if [ "$1" = "test" ]; then
+    shift
+    docker stop reactive-playground
+    docker rm reactive-playground
 
-docker run -it --name reactive-playground \
- -h reactive \
- -v ~/.m2:/home/user/.m2 \
- -p 8181:8181 \
- -p 15005:5005 \
- mhus/reactive-playground:7.0.0-SNAPSHOT debug
+    docker run -it --name reactive-playground \
+     -h reactive \
+     -v ~/.m2:/home/user/.m2 \
+     -p 8181:8181 \
+     -p 15005:5005 \
+     mhus/reactive-playground:7.0.0-SNAPSHOT debug
+fi
 
-#kubectl apply -f kubernetes/reactive-test.yaml
-
+if [ "$1" = "k8s" ]; then
+    shift
+    kubectl apply -f kubernetes/reactive-test.yaml
+fi
